Refactor parts of GitHubIntegration into BaseIntegration #346
