+++ workflow/integrations/github.py
@@ -8,33 +8,18 @@
 
 from urllib.parse import urlparse
 
-from django.conf import settings
-
-import requests
+from workflow.integrations import BaseIntegration
 
 GITHUB_API_URL = "https://api.github.com"
-DEJACODE_SITE_URL = settings.SITE_URL.rstrip("/")
 
 
-class GitHubIntegration:
+class GitHubIntegration(BaseIntegration):
     """
     A class for managing GitHub issue creation, updates, and comments
     from DejaCode requests.
     """
 
     api_url = GITHUB_API_URL
-    default_timeout = 10
-
-    def __init__(self, dataspace):
-        if not dataspace:
-            raise ValueError("Dataspace must be provided.")
-        self.dataspace = dataspace
-        self.session = self.get_session()
-
-    def get_session(self):
-        session = requests.Session()
-        session.headers.update(self.get_headers())
-        return session
 
     def get_headers(self):
         github_token = self.dataspace.get_configuration(field_name="github_token")
@@ -137,40 +122,3 @@ class GitHubIntegration:
             raise ValueError("Incomplete GitHub repository path.")
 
         return f"{path_parts[0]}/{path_parts[1]}"
-
-    @staticmethod
-    def make_issue_title(request):
-        return f"[DEJACODE] {request.title}"
-
-    @staticmethod
-    def make_issue_body(request):
-        request_url = f"{DEJACODE_SITE_URL}{request.get_absolute_url()}"
-        label_fields = [
-            ("📝 Request Template", request.request_template),
-            ("📦 Product Context", request.product_context),
-            ("📌 Applies To", request.content_object),
-            ("🙋 Submitted By", request.requester),
-            ("👤 Assigned To", request.assignee),
-            ("🚨 Priority", request.priority),
-            ("🗒️ Notes", request.notes),
-            ("🔗️ DejaCode URL", request_url),
-        ]
-
-        lines = []
-        for label, value in label_fields:
-            if value:
-                lines.append(f"### {label}\n{value}")
-
-        lines.append("----")
-
-        for question in request.get_serialized_data_as_list():
-            label = question.get("label")
-            value = question.get("value")
-            input_type = question.get("input_type")
-
-            if input_type == "BooleanField":
-                value = "Yes" if str(value).lower() in ("1", "true", "yes") else "No"
-
-            lines.append(f"### {label}\n{value}")
-
-        return "\n\n".join(lines)
